Allow decimal center and radius

+++ bukkit/src/main/java/org/popcraft/chunky/platform/BukkitConfig.java
@@ -34,11 +34,11 @@ public class BukkitConfig implements Config {
         if (config.getBoolean(world_key + "cancelled", false)) {
             return Optional.empty();
         }
-        int radiusX = config.getInt(world_key + "radius", 500);
-        int radiusZ = config.getInt(world_key + "z-radius", radiusX);
+        double radiusX = config.getDouble(world_key + "radius", 500);
+        double radiusZ = config.getDouble(world_key + "z-radius", radiusX);
         Selection.Builder selection = Selection.builder(world)
-                .centerX(config.getInt(world_key + "x-center", 0))
-                .centerZ(config.getInt(world_key + "z-center", 0))
+                .centerX(config.getDouble(world_key + "x-center", 0))
+                .centerZ(config.getDouble(world_key + "z-center", 0))
                 .radiusX(radiusX)
                 .radiusZ(radiusZ)
                 .pattern(config.getString(world_key + "iterator", "loop"))
